requirements.txt
Flask==1.0.2
werkzeug==0.15.5
numpy==1.16.1
requests==2.22.0
Pillow==6.1.0
